requirements.txt
numpy
tensorflow-gpu==1.14
matplotlib
nilearn
scipy
sklearn
batchgenerators